TODO: Axios uses xhr instead of http backend. Very unstable hacky solution here as backup.
